split data processing for training to new file
